Migrate to tree implementation
This changes the implementation to no longer return Attrs that have a
slice of keys and a value, but to give keys a `Children` property so
that they can be traversed in a tree-like fashion.

This better aligns with how structured logs work conceptually and
still provides an easier-to-use interface over the default slog
Handler interface for a minor performance penalty.